backend/social_media_manager/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==40.0.2

# OAuth & API Integration
requests==2.29.0
requests-oauthlib==1.3.1
oauthlib==3.2.2

# Scheduling
apscheduler==3.10.1
celery==5.2.7
redis==4.5.5

# Utilities
pydantic[email]==1.10.7
python-dateutil==2.8.2

# Optional Elasticsearch integration
elasticsearch==8.7.0